# Optimize memory footprint of pdf problem transcription task