--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -45,9 +45,10 @@ jobs:
 
       - uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065 # v5
         with:
-          python-version: "3.12"
+          python-version-file: "pyproject.toml"
           cache: "poetry"
-
+      - name: Check Python version
+        run: python --version
       - name: Validate lockfile
         run: poetry check --lock