--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -47,6 +47,8 @@ jobs:
         with:
           python-version-file: "pyproject.toml"
           cache: "poetry"
+      - name: Install poetry with pip
+        run: python -m pip install poetry
       - name: Check Python version
         run: python --version
       - name: Validate lockfile
@@ -54,7 +56,9 @@ jobs:
       - name: Set Poetry Python
         run: poetry env use python3.12
       - name: Install dependencies
-        run: poetry install --no-interaction
+        run: |
+          source $(poetry env info --path)/bin/activate
+          poetry install --no-interaction
 
       - name: Create test local state
         run: ./scripts/test/stub-data.sh