--- a/pyproject.toml
+++ b/pyproject.toml
@@ -111,7 +111,6 @@ uwsgi = "^2.0.29"
 uwsgitop = "^0.12"
 wrapt = "^1.14.1"
 youtube-transcript-api = "^1.0.0"
-pdf2image = "^1.17.0"
 pypdfium2 = "^4.30.0"
 
 [tool.poetry.group.dev.dependencies]