--- a/pyproject.toml
+++ b/pyproject.toml
@@ -34,7 +34,7 @@ django-cors-headers = "^4.0.0"
 django-filter = "^2.4.0"
 django-guardian = "^3.0.0"
 django-health-check = { git = "https://github.com/revsys/django-health-check", rev="b0500d14c338040984f02ee34ffbe6643b005084" }  # pragma: allowlist secret
-pypdfium2 = { git = "https://github.com/pypdfium2-team/pypdfium2"}  # pragma: allowlist secret
+pypdfium2 = { git = "https://github.com/pypdfium2-team/pypdfium2"}
 
 django-imagekit = "^5.0.0"
 django-ipware = "^7.0.0"